backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0

# Computer Vision
ultralytics==8.1.0
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# OCR
easyocr==1.7.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Email (Brevo Transactional API)
sib-api-v3-sdk==7.6.0
